requirements.txt

Flask
yfinance
python-dotenv
google-generativeai
langchain
langchain_groq
